Addressed comments made on pr
